Split MatchStudent to MatchStudentDefault and MatchStudentChoiceReuse
MatchStudentDefault reverted drop() function to original behavior
MatchStudentChoiceReuse overrides drop() function to allow reusing
choice
Add choiceReuseEnabled field to MatchContent
