+++ src/assets/wise5/components/match/MatchContent.ts
@@ -4,6 +4,7 @@ import { Choice } from './choice';
 export interface MatchContent extends ComponentContent {
   buckets: any[];
   canCreateChoices: boolean;
+  choiceReuseEnabled: boolean;
   choices: Choice[];
   feedback: any[];
   ordered: boolean;
