+++ src/assets/wise5/components/match/match-student/MatchCdkDragDrop.ts
@@ -1,0 +1,4 @@
+import { CdkDragDrop } from '@angular/cdk/drag-drop';
+
+export interface MatchCdkDragDrop<Container, Item>
+  extends CdkDragDrop<Container, Container, Item> {}
