+++ src/assets/wise5/components/match/match-student/match-student.module.ts
@@ -4,9 +4,17 @@ import { AddMatchChoiceDialog } from './add-match-choice-dialog/add-match-choice
 import { MatchStudent } from './match-student.component';
 import { StudentComponentModule } from '../../../../../app/student/student.component.module';
 import { MatchCommonModule } from '../match-common.module';
+import { MatchStudentChoiceReuse } from './match-student-choice-reuse/match-student-choice-reuse';
+import { MatchStudentDefault } from './match-student-default/match-student-default.component';
 
 @NgModule({
-  declarations: [AddChoiceButton, AddMatchChoiceDialog, MatchStudent],
+  declarations: [
+    AddChoiceButton,
+    AddMatchChoiceDialog,
+    MatchStudent,
+    MatchStudentDefault,
+    MatchStudentChoiceReuse
+  ],
   imports: [MatchCommonModule, StudentComponentModule],
   exports: [MatchStudent]
 })
